% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmatransport.R
\docType{package}
\name{nmatransport-package}
\alias{nmatransport}
\alias{nmatransport-package}
\title{nmatransport: Population Transportability and Transitivity Assessment in Network Meta-Analysis}
\description{
Extends standard network meta-analysis (NMA) to target populations by adjusting for differences in effect modifiers across the network. Implements matching-adjusted techniques for aggregate data NMA, automated transitivity assessments, and personalized treatment rankings to solve generalizability gaps in evidence synthesis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mahmood726-cyber/nmatransport}
  \item Report bugs at \url{https://github.com/mahmood726-cyber/nmatransport/issues}
}

}
\author{
\strong{Maintainer}: Mahmood Ahmad \email{mahmood726@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

}
\keyword{internal}
