% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmatransport.R
\name{print.transported_nma}
\alias{print.transported_nma}
\title{Print Transported NMA Results}
\usage{
\method{print}{transported_nma}(x, ...)
}
\arguments{
\item{x}{transported_nma object}

\item{...}{Additional arguments}
}
\description{
Print Transported NMA Results
}
